# server/ee/libs/embedded/embedded-configuration/embedded-configuration-rest/embedded-configuration-rest-api/generated/src/main/java/com/bytechef/ee/embedded/configuration/web/rest/ComponentDefinitionApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.24.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bytechef.ee.embedded.configuration.web.rest;

import com.bytechef.ee.embedded.configuration.web.rest.model.ComponentDefinitionBasicModel;
import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-31T17:46:26.627594+02:00[Europe/Zagreb]", comments = "Generator version: 7.24.0")
@Validated
@Tag(name = "component-definition", description = "The Embedded Component Definition Internal API")
public interface ComponentDefinitionApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_COMPONENT_DEFINITIONS = "/component-definitions";
    /**
     * GET /component-definitions : Get all component definitions
     * Get all component definitions.
     *
     * @param actionDefinitions Use for filtering components which define action definitions. (optional)
     * @param clusterElementDefinitions Use for filtering components which define cluster element definitions. (optional)
     * @param connectionDefinitions Use for filtering components which define connection definitions. (optional)
     * @param triggerDefinitions Use for filtering components which define trigger definitions. (optional)
     * @param include The list of component names to include in the result. (optional)
     * @return Successful operation. (status code 200)
     */
    @Operation(
        operationId = "getComponentDefinitions",
        summary = "Get all component definitions",
        description = "Get all component definitions.",
        tags = { "component-definition" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ComponentDefinitionBasicModel.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ComponentDefinitionApi.PATH_GET_COMPONENT_DEFINITIONS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<ComponentDefinitionBasicModel>> getComponentDefinitions(
        @Parameter(name = "actionDefinitions", description = "Use for filtering components which define action definitions.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "actionDefinitions", required = false) @Nullable Boolean actionDefinitions,
        @Parameter(name = "clusterElementDefinitions", description = "Use for filtering components which define cluster element definitions.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "clusterElementDefinitions", required = false) @Nullable Boolean clusterElementDefinitions,
        @Parameter(name = "connectionDefinitions", description = "Use for filtering components which define connection definitions.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "connectionDefinitions", required = false) @Nullable Boolean connectionDefinitions,
        @Parameter(name = "triggerDefinitions", description = "Use for filtering components which define trigger definitions.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "triggerDefinitions", required = false) @Nullable Boolean triggerDefinitions,
        @Parameter(name = "include", description = "The list of component names to include in the result.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "include", required = false) @Nullable List<String> include
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"actionsCount\" : 0, \"clusterElementsCount\" : { \"key\" : 6 }, \"clusterRoot\" : true, \"componentCategories\" : [ { \"name\" : \"name\", \"label\" : \"label\" }, { \"name\" : \"name\", \"label\" : \"label\" } ], \"description\" : \"description\", \"icon\" : \"icon\", \"inputsCount\" : 1, \"name\" : \"name\", \"title\" : \"title\", \"triggersCount\" : 5, \"version\" : 5 }, { \"actionsCount\" : 0, \"clusterElementsCount\" : { \"key\" : 6 }, \"clusterRoot\" : true, \"componentCategories\" : [ { \"name\" : \"name\", \"label\" : \"label\" }, { \"name\" : \"name\", \"label\" : \"label\" } ], \"description\" : \"description\", \"icon\" : \"icon\", \"inputsCount\" : 1, \"name\" : \"name\", \"title\" : \"title\", \"triggersCount\" : 5, \"version\" : 5 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}