# server/ee/libs/embedded/embedded-configuration/embedded-configuration-rest/embedded-configuration-rest-api/generated/src/main/java/com/bytechef/ee/embedded/configuration/web/rest/IntegrationInstanceApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.24.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bytechef.ee.embedded.configuration.web.rest;

import com.bytechef.ee.embedded.configuration.web.rest.model.IntegrationInstanceModel;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-31T17:46:26.627594+02:00[Europe/Zagreb]", comments = "Generator version: 7.24.0")
@Validated
@Tag(name = "integration-instance", description = "the integration-instance API")
public interface IntegrationInstanceApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_DELETE_INTEGRATION_INSTANCE = "/integration-instances/{id}";
    /**
     * DELETE /integration-instances/{id} : Delete an integration instance
     * Delete an integration instance.
     *
     * @param id The id of an integration instance. (required)
     * @return Successful operation. (status code 204)
     */
    @Operation(
        operationId = "deleteIntegrationInstance",
        summary = "Delete an integration instance",
        description = "Delete an integration instance.",
        tags = { "integration-instance" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Successful operation.")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = IntegrationInstanceApi.PATH_DELETE_INTEGRATION_INSTANCE
    )
    default ResponseEntity<Void> deleteIntegrationInstance(
        @Parameter(name = "id", description = "The id of an integration instance.", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_ENABLE_INTEGRATION_INSTANCE = "/integration-instances/{id}/enable/{enable}";
    /**
     * PATCH /integration-instances/{id}/enable/{enable} : Enable/disable an integration instance
     * Enable/disable an integration instance.
     *
     * @param id The id of an integration instance. (required)
     * @param enable Enable/disable the integration instance. (required)
     * @return Successful operation. (status code 204)
     */
    @Operation(
        operationId = "enableIntegrationInstance",
        summary = "Enable/disable an integration instance",
        description = "Enable/disable an integration instance.",
        tags = { "integration-instance" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Successful operation.")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = IntegrationInstanceApi.PATH_ENABLE_INTEGRATION_INSTANCE
    )
    default ResponseEntity<Void> enableIntegrationInstance(
        @Parameter(name = "id", description = "The id of an integration instance.", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "enable", description = "Enable/disable the integration instance.", required = true, in = ParameterIn.PATH) @PathVariable("enable") Boolean enable
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_ENABLE_INTEGRATION_INSTANCE_WORKFLOW = "/integration-instances/{id}/workflows/{workflowId}/enable/{enable}";
    /**
     * PATCH /integration-instances/{id}/workflows/{workflowId}/enable/{enable} : Enable/disable an integration instance workflow
     * Enable/disable an integration instance workflow.
     *
     * @param id The id of an integration instance. (required)
     * @param workflowId The id of an integration instance workflow. (required)
     * @param enable Enable/disable the integration instance workflow. (required)
     * @return Successful operation. (status code 204)
     */
    @Operation(
        operationId = "enableIntegrationInstanceWorkflow",
        summary = "Enable/disable an integration instance workflow",
        description = "Enable/disable an integration instance workflow.",
        tags = { "integration-instance" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Successful operation.")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = IntegrationInstanceApi.PATH_ENABLE_INTEGRATION_INSTANCE_WORKFLOW
    )
    default ResponseEntity<Void> enableIntegrationInstanceWorkflow(
        @Parameter(name = "id", description = "The id of an integration instance.", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "workflowId", description = "The id of an integration instance workflow.", required = true, in = ParameterIn.PATH) @PathVariable("workflowId") String workflowId,
        @Parameter(name = "enable", description = "Enable/disable the integration instance workflow.", required = true, in = ParameterIn.PATH) @PathVariable("enable") Boolean enable
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_INTEGRATION_INSTANCE = "/integration-instances/{id}";
    /**
     * GET /integration-instances/{id} : Get an integration instance by id
     * Get an integration instance by id.
     *
     * @param id The id of an integration instance. (required)
     * @return The integration instance object. (status code 200)
     */
    @Operation(
        operationId = "getIntegrationInstance",
        summary = "Get an integration instance by id",
        description = "Get an integration instance by id.",
        tags = { "integration-instance" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The integration instance object.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = IntegrationInstanceModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = IntegrationInstanceApi.PATH_GET_INTEGRATION_INSTANCE,
        produces = { "application/json" }
    )
    default ResponseEntity<IntegrationInstanceModel> getIntegrationInstance(
        @Parameter(name = "id", description = "The id of an integration instance.", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"connectionId\" : 0, \"connectedUserId\" : 6, \"createdBy\" : \"createdBy\", \"createdDate\" : \"2000-01-23T04:56:07.000+00:00\", \"enabled\" : true, \"environmentId\" : 1, \"id\" : 5, \"lastExecutionDate\" : \"2000-01-23T04:56:07.000+00:00\", \"lastModifiedBy\" : \"lastModifiedBy\", \"lastModifiedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"integrationInstanceConfigurationId\" : 5, \"integrationInstanceConfiguration\" : { \"createdBy\" : \"createdBy\", \"createdDate\" : \"2000-01-23T04:56:07.000+00:00\", \"description\" : \"description\", \"enabled\" : true, \"environmentId\" : 2, \"id\" : 7, \"integrationId\" : 9, \"integrationVersion\" : 3, \"lastModifiedBy\" : \"lastModifiedBy\", \"lastModifiedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"name\" : \"name\", \"authorizationType\" : \"API_KEY\" }, \"integrationInstanceWorkflows\" : [ { \"createdBy\" : \"createdBy\", \"createdDate\" : \"2000-01-23T04:56:07.000+00:00\", \"inputs\" : { \"key\" : \"\" }, \"enabled\" : true, \"id\" : 2, \"integrationInstanceConfigurationWorkflowId\" : 4, \"lastModifiedBy\" : \"lastModifiedBy\", \"lastModifiedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"workflowId\" : \"workflowId\" }, { \"createdBy\" : \"createdBy\", \"createdDate\" : \"2000-01-23T04:56:07.000+00:00\", \"inputs\" : { \"key\" : \"\" }, \"enabled\" : true, \"id\" : 2, \"integrationInstanceConfigurationWorkflowId\" : 4, \"lastModifiedBy\" : \"lastModifiedBy\", \"lastModifiedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"workflowId\" : \"workflowId\" } ], \"__version\" : 7 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}