# server/ee/libs/embedded/embedded-configuration/embedded-configuration-rest/embedded-configuration-rest-impl/openapi.yaml
...
openapi: "3.0.1"
info:
  title: "The Embedded Configuration Internal API"
  version: "1"
servers:
  - url: "/api/embedded/internal"
...
paths:
  /component-definitions:
    get:
      description: "Get all component definitions."
      summary: "Get all component definitions"
      tags:
        - "component-definition"
      operationId: "getComponentDefinitions"
      parameters:
        - name: "actionDefinitions"
          description: "Use for filtering components which define action definitions."
          in: "query"
          required: false
          schema:
            type: "boolean"
        - name: "clusterElementDefinitions"
          description: "Use for filtering components which define cluster element definitions."
          in: "query"
          required: false
          schema:
            type: "boolean"
        - name: "connectionDefinitions"
          description: "Use for filtering components which define connection definitions."
          in: "query"
          required: false
          schema:
            type: "boolean"
        - name: "triggerDefinitions"
          description: "Use for filtering components which define trigger definitions."
          in: "query"
          required: false
          schema:
            type: "boolean"
        - name: "include"
          description: "The list of component names to include in the result."
          in: "query"
          required: false
          schema:
            type: "array"
            items:
              type: "string"
      responses:
        "200":
          description: "Successful operation."
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "../../../../../../libs/platform/platform-configuration/platform-configuration-rest/platform-configuration-rest-impl/openapi.yaml/#/components/schemas/ComponentDefinitionBasic"
...
  /integration-instances/{id}:
    get:
      description: "Get an integration instance by id."
      summary: "Get an integration instance by id"
      tags:
        - "integration-instance"
      operationId: "getIntegrationInstance"
      parameters:
        - name: "id"
          description: "The id of an integration instance."
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "The integration instance object."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationInstance"
    delete:
      description: "Delete an integration instance."
      summary: "Delete an integration instance"
      tags:
        - "integration-instance"
      operationId: "deleteIntegrationInstance"
      parameters:
        - name: "id"
          description: "The id of an integration instance."
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "204":
          description: "Successful operation."
  /integration-instances/{id}/enable/{enable}:
    patch:
      description: "Enable/disable an integration instance."
      summary: "Enable/disable an integration instance"
      tags:
        - "integration-instance"
      operationId: "enableIntegrationInstance"
      parameters:
        - name: "id"
          description: "The id of an integration instance."
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "enable"
          description: "Enable/disable the integration instance."
          in: "path"
          required: true
          schema:
            type: "boolean"
      responses:
        "204":
          description: "Successful operation."
  /integration-instances/{id}/workflows/{workflowId}/enable/{enable}:
    patch:
      description: "Enable/disable an integration instance workflow."
      summary: "Enable/disable an integration instance workflow"
      tags:
        - "integration-instance"
      operationId: "enableIntegrationInstanceWorkflow"
      parameters:
        - name: "id"
          description: "The id of an integration instance."
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "workflowId"
          description: "The id of an integration instance workflow."
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "enable"
          description: "Enable/disable the integration instance workflow."
          in: "path"
          required: true
          schema:
            type: "boolean"
      responses:
        "204":
          description: "Successful operation."
...
components:
  schemas:
...
    IntegrationInstance:
      description: "Contains configurations and connections required for the execution\
        \ of integration workflows for a connected user."
      allOf:
        - $ref: "#/components/schemas/IntegrationInstanceBasic"
        - type: "object"
          required:
            - "connectionId"
            - "name"
          properties:
            integrationInstanceConfiguration:
              $ref: "#/components/schemas/IntegrationInstanceConfigurationBasic"
            integrationInstanceWorkflows:
              description: "The array of integration instance workflows."
              type: "array"
              items:
                $ref: "#/components/schemas/IntegrationInstanceWorkflow"
              readOnly: true
            __version:
              type: "integer"
    IntegrationInstanceBasic:
      description: "Contains configurations and connections required for the execution\
        \ of integration workflows for a connected user."
      type: "object"
      required:
        - "connectionId"
        - "name"
      properties:
        connectionId:
          description: "The id of a connection."
          type: "integer"
          format: "int64"
          readOnly: true
        connectedUserId:
          description: "The id of a connected user."
          type: "integer"
          format: "int64"
          readOnly: true
        createdBy:
          description: "The created by."
          type: "string"
          readOnly: true
        createdDate:
          description: "The created date."
          type: "string"
          format: "date-time"
          readOnly: true
        enabled:
          description: "If an integration instance is enabled or not."
          type: "boolean"
        environmentId:
          description: "The id of an environment."
          type: "integer"
          format: "int64"
        id:
          description: "The id of an integration instance."
          type: "integer"
          format: "int64"
          readOnly: true
        lastExecutionDate:
          description: "The last execution date."
          type: "string"
          format: "date-time"
          readOnly: true
        lastModifiedBy:
          description: "The last modified by."
          type: "string"
          readOnly: true
        lastModifiedDate:
          description: "The last modified date."
          type: "string"
          format: "date-time"
          readOnly: true
        integrationInstanceConfigurationId:
          description: "Th id of an integration instance configuration."
          type: "integer"
          format: "int64"
...
    IntegrationInstanceConfigurationBasic:
      description: "Contains configurations and connections required for the execution\
        \ of integration workflows."
      type: "object"
      required:
        - "name"
      properties:
        createdBy:
          description: "The created by."
          type: "string"
          readOnly: true
        createdDate:
          description: "The created date."
          type: "string"
          format: "date-time"
          readOnly: true
        description:
          description: "The description of an integration configuration."
          type: "string"
        enabled:
          description: "If an integration instance configuration is enabled or not."
          type: "boolean"
        environmentId:
          description: "The id of an environment."
          type: "integer"
          format: "int64"
        id:
          description: "The id of an integration instance configuration."
          type: "integer"
          format: "int64"
          readOnly: true
        integrationId:
          description: "Th id of an integration."
          type: "integer"
          format: "int64"
        integrationVersion:
          description: "The version of an integration."
          type: "integer"
        lastModifiedBy:
          description: "The last modified by."
          type: "string"
          readOnly: true
        lastModifiedDate:
          description: "The last modified date."
          type: "string"
          format: "date-time"
          readOnly: true
        name:
          description: "The name of an integration instance configuration."
          type: "string"
        authorizationType:
          $ref: "../../../../../../libs/platform/platform-configuration/platform-configuration-rest/platform-configuration-rest-impl/openapi.yaml#/components/schemas/AuthorizationType"
...
    IntegrationInstanceWorkflow:
      description: "Contains user configurations for the execution\
        \ of a particular integration workflow."
      type: "object"
      required:
        - "enabled"
        - "integrationInstanceConfigurationWorkflowId"
        - "workflowId"
      properties:
        createdBy:
          description: "The created by."
          type: "string"
          readOnly: true
        createdDate:
          description: "The created date."
          type: "string"
          format: "date-time"
          readOnly: true
        inputs:
          description: "The input parameters of an integration instance used as workflow\
            \ input values."
          type: "object"
          additionalProperties: true
        enabled:
          description: "If a workflow is enabled or not in the integration instance workflow."
          type: "boolean"
        id:
          description: "The id of a project deployment workflow."
          type: "integer"
          format: "int64"
          readOnly: true
        integrationInstanceConfigurationWorkflowId:
          description: "The id of a integration instance configuration workflow."
          type: "integer"
          format: "int64"
        lastModifiedBy:
          description: "The last modified by."
          type: "string"
          readOnly: true
        lastModifiedDate:
          description: "The last modified date."
          type: "string"
          format: "date-time"
          readOnly: true
        workflowId:
          description: "The id of a workflow."
          type: "string"
...